+++ hax-lib/core-models/src/core/ops.rs
@@ -16,21 +16,40 @@ pub mod arith {
         fn div(self, rhs: Rhs) -> Self::Output;
     }
     pub trait AddAssign<Rhs = Self> {
-        type Output;
-        fn add_assign(self, rhs: Rhs) -> Self::Output;
+        fn add_assign(&mut self, rhs: Rhs);
     }
     pub trait SubAssign<Rhs = Self> {
-        type Output;
-        fn sub_assign(self, rhs: Rhs) -> Self::Output;
+        fn sub_assign(&mut self, rhs: Rhs);
     }
     pub trait MulAssign<Rhs = Self> {
-        type Output;
-        fn mul_assign(self, rhs: Rhs) -> Self::Output;
+        fn mul_assign(&mut self, rhs: Rhs);
     }
     pub trait DivAssign<Rhs = Self> {
-        type Output;
-        fn div_assign(self, rhs: Rhs) -> Self::Output;
+        fn div_assign(&mut self, rhs: Rhs);
     }
+
+    macro_rules! int_trait_impls {
+        ($($Self:ty)*) => {
+            use hax_lib::ToInt;
+            $(
+            #[hax_lib::attributes]
+            impl crate::ops::arith::AddAssign<$Self> for $Self {
+                #[hax_lib::requires(self.to_int() + rhs.to_int() <= $Self::MAX.to_int())]
+                fn add_assign(&mut self, rhs: $Self) {
+                    *self = *self + rhs
+                }
+            }
+            #[hax_lib::attributes]
+            impl crate::ops::arith::SubAssign<$Self> for $Self {
+                #[hax_lib::requires(self.to_int() - rhs.to_int() >= 0.to_int())]
+                fn sub_assign(&mut self, rhs: $Self) {
+                    *self = *self - rhs
+                }
+            })*
+        }
+    }
+
+    int_trait_impls!(u8 u16 u32 u64);
 }
 
 pub mod bit {
@@ -56,8 +75,6 @@ pub mod control_flow {
 }
 
 pub mod index {
-    // TODO remove
-    #[hax_lib::fstar::before("open Rust_primitives.Integers")]
     trait Index<Idx> {
         type Output;
         fn index(&self, i: Idx) -> &Self::Output;
@@ -65,25 +82,6 @@ pub mod index {
 }
 
 pub mod function {
-    /* These instances provide implementations of the F* type classes corresponding to Fn traits for anonymous functions.
-    This ensures that passing a closure where something implementing Fn works when translated to F* */
-    #[hax_lib::fstar::after(
-        "unfold instance fnonce_arrow t u
-  : t_FnOnce (t -> u) t = {
-    f_Output = u;
-    f_call_once_pre = (fun _ _ -> true);
-    f_call_once_post = (fun (x0: t -> u) (x1: t) (res: u) -> res == x0 x1);
-    f_call_once = (fun (x0: t -> u) (x1: t) -> x0 x1);
-  } 
-unfold instance fnonce_arrow_binder t u
-  : t_FnOnce (_:t -> u) t = {
-    f_Output = u;
-    f_call_once_pre = (fun _ _ -> true);
-    f_call_once_post = (fun (x0: (_:t -> u)) (x1: t) (res: u) -> res == x0 x1);
-    f_call_once = (fun (x0: (_:t -> u)) (x1: t) -> x0 x1);
-  }
-"
-    )]
     #[hax_lib::attributes]
     pub trait FnOnce<Args> {
         type Output;
@@ -94,6 +92,24 @@ unfold instance fnonce_arrow_binder t u
     pub trait Fn<Args>: FnOnce<Args> {
         #[hax_lib::requires(true)]
         fn call(&self, args: Args) -> Self::Output;
+    }
+
+    /* These instances provide implementations of the F* type classes corresponding to Fn traits for anonymous functions.
+    This ensures that passing a closure where something implementing Fn works when translated to F* */
+    #[hax_lib::fstar::after(
+        "unfold instance fnonce_arrow_binder t u
+  : t_FnOnce (_:t -> u) t = {
+    f_Output = u;
+    f_call_once_pre = (fun _ _ -> true);
+    f_call_once_post = (fun (x0: (_:t -> u)) (x1: t) (res: u) -> res == x0 x1);
+    f_call_once = (fun (x0: (_:t -> u)) (x1: t) -> x0 x1);
+  }"
+    )]
+    impl<Args, Out> FnOnce<Args> for fn(Args) -> Out {
+        type Output = Out;
+        fn call_once(&self, args: Args) -> Out {
+            self(args)
+        }
     }
 }